Refactored from test_uint256 to test_convert_uint256
